requirements.txt
# Kalshi Trends - Data Pipeline & Analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.15.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
requests>=2.31.0
httpx>=0.25.0
python-dotenv>=1.0.0
pyarrow>=14.0.0
streamlit>=1.28.0
